Move LLVMTypeHelper to allow usage of old and new types
